Remove ComplexProduct
The former `compcomp_t` and then `ComplexProduct` was just a half-assed
approach to treat analytically vanishing imaginary parts of certain
traces. The user still had to figure out which columns to take depending
on analytic derivations. This does not scale for more than two traces.

Therefore we remove this stuff at this point and might add some fancy
logic later on such that pure noise contributions are discarded
automatically in the right spots.