Let the result be `sum_ - c_`
Querying for the result must be the same as doing `+= 0.0`. If we do the
latter, then we have `y = - c_` and then `t = sum_ + y` and `t` becomes
the new sum. In that case we have `sum_ = sum_ + (0.0 - c_)` and
therefore the end result should be `sum_ - c_` I believe.